refactor: 사다리 타기 이동 부분 추가로 Controller, Service 로 구분
