+++ src/main/java/nextstep/ladder/Service/LadderService.java
@@ -1,0 +1,12 @@
+package nextstep.ladder.Service;
+
+import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.domain.LadderRewards;
+import nextstep.ladder.domain.Participants;
+
+public class LadderService {
+
+    public static void run(Participants participants, Ladder ladder, LadderRewards ladderRewards) {
+
+    }
+}
